Move some structs to their own modules
All logic remains unchanged, this commit only moves code around.
